fixing the vite polyfill configuration 🃏
