+++ packages/vite/.gitignore
@@ -23,3 +23,4 @@ dist-ssr
 *.sln
 *.sw?
 .vercel
+.yarn/
