+++ packages/vite/src/App.tsx
@@ -15,9 +15,7 @@ const client = new ApolloClient({
   cache: new InMemoryCache(),
 })
 
-const App: React.FC = () => {
-  if(true) return <>Testing</>
-  return (
+const App: React.FC = () => (
   <ChakraProvider>
     <Helmet>
       <link
@@ -35,7 +33,7 @@ const App: React.FC = () => {
        {/* <Component {...pageProps} /> */}
       </Web3ContextProvider>
     </ApolloProvider>
-  </ChakraProvider>)
-}
+  </ChakraProvider>
+)
 
 export default App
